requirements.txt
langgraph>=0.2.20,<0.3
fastapi==0.100.0
uvicorn==0.22.0
yfinance==0.2.53
pandas==2.1.0
python-dotenv==1.0.0
pydantic==2.10.6
langchain==0.3.18
langchain_google_genai==2.0.9
requests==2.31.0
requests_cache